Help channels: rename modules to use singular tense
Plural names imply the modules contain homogenous content. For example,
"channels" implies the module contains multiple kinds of channels.
